+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 
 * Update the `org.realityforge.braincheck` artifacts to version `1.32.0`.
+* Add the `sting.warnings_as_errors` annotation processor option to upgrade Sting warnings to errors.
 * Introduce the `@sting.Factory` annotation to generate injectable factory implementations that can be included in
   injectors via `@StingProvider`.
 * Update framework integration documentation and diagnostics to clarify that